services/runner/requirements.txt
openai==1.12.0
anthropic==0.18.1
google-generativeai==0.3.2
httpx==0.26.0
pydantic==2.5.3
redis==5.0.1
asyncpg==0.29.0
python-dotenv==1.0.0
tenacity==8.2.3
structlog==24.1.0
pytest==7.4.4
pytest-asyncio==0.23.3
black==23.12.1
ruff==0.1.11
beautifulsoup4==4.12.0
lxml==4.9.0
aiohttp==3.9.0